update at 12:32 pm
renamed a couple of files to make more sense. made a party game: player gets 6 blobs of texts and goes through the portal. after the portal the player goes to a void to read text. Not linked but after it goes to a hospital. After leaving the hospital the player goes to the outside world. In the outside world the player can only walk on the road or sidewalk and after 10 seconds they get hit by a car.